+++ src/libs/vector2.js
@@ -2,10 +2,17 @@ export class Vector2 {
     x;
     y;
     
+    /**
+     * Returns a zero vector
+     */
     static get zero() {
-        return new this();
+        return new this(0, 0);
     }
 
+    /**
+     * Add's another vector to this vector
+     * @param {Vector2} otherVec 
+     */
     add(otherVec) {
         this.x += otherVec.x;
         this.y += otherVec.y;